Revert to 56e05fced

+++ pyfr/quadrules/__init__.py
@@ -38,8 +38,8 @@ class BaseTabulatedQuadRule(object):
             pts = [p[0] for p in pts]
 
         # Cast
-        self.pts = np.array(pts)
-        self.wts = np.array(wts)
+        self.pts = np.array(pts, dtype=np.float)
+        self.wts = np.array(wts, dtype=np.float)
 
 
 class BaseStoredQuadRule(BaseTabulatedQuadRule):
